Calculate vulnerability

Calculate vulnerability

calc_topsis(
  indicator.name.sens,
  indicator.name.exp,
  indicator.name.adap,
  cost.data.comms,
  file.ind
)

Arguments

file.ind

Examples

calc_topsis(indicator.name.sens="sensitivity", indicator.name.exp= "exposure_historical", indicator.name.adap= "adaptive_capacity", cost.data.comms, file.ind = "historical")
#> Rows: 2219 Columns: 16
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (3): CVE_MUN, CVE_LOC, component
#> dbl (13): ua_area_loc_kmsq, tot_denue_aqua, aqua_pcapita, prop_ueaqua, tot_d...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2219 Columns: 19
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (3): CVE_MUN, CVE_LOC, component
#> dbl (16): cbpm_mean_historical, observed_dissolved_o2_mean_historical, world...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 2219 Columns: 83
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr  (3): CVE_MUN, CVE_LOC, component
#> dbl (80): acc_1health, acc_2health, acc_3health, acc_presc, acc_elem, acc_se...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> [1] 1
#> [1] "04008"
#> Error in select(., CVE_MUN, CVE_LOC): could not find function "select"